+++ core/amber/src/main/protobuf/edu/uci/ics/amber/engine/architecture/worker.proto
