+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/deploysemantics/layer/WorkerInfo.scala
@@ -1,7 +1,7 @@
 package edu.uci.ics.amber.engine.architecture.deploysemantics.layer
 
 import edu.uci.ics.amber.engine.architecture.worker.WorkerStatistics
-import edu.uci.ics.amber.engine.common.statetransition.WorkerState
 import edu.uci.ics.amber.engine.common.virtualidentity.ActorVirtualIdentity
+import edu.uci.ics.amber.engine.common.worker.WorkerState
 
 case class WorkerInfo(id: ActorVirtualIdentity, var state: WorkerState, var stats: WorkerStatistics)
