+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/WorkerStatistics.scala
@@ -1,7 +1,7 @@
 package edu.uci.ics.amber.engine.architecture.worker
 
-import edu.uci.ics.amber.engine.common.statetransition.WorkerState
 import edu.uci.ics.amber.engine.common.tuple.ITuple
+import edu.uci.ics.amber.engine.common.worker.WorkerState
 import edu.uci.ics.texera.workflow.common.IncrementalOutputMode
 
 case class WorkerStatistics(
