+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/worker/promisehandlers/AddPartitioningHandler.scala
@@ -4,8 +4,8 @@ import edu.uci.ics.amber.engine.architecture.sendsemantics.partitionings.Partiti
 import edu.uci.ics.amber.engine.architecture.worker.WorkerAsyncRPCHandlerInitializer
 import edu.uci.ics.amber.engine.architecture.worker.promisehandlers.AddPartitioningHandler.AddPartitioning
 import edu.uci.ics.amber.engine.common.rpc.AsyncRPCServer.{CommandCompleted, ControlCommand}
-import edu.uci.ics.amber.engine.common.statetransition.WorkerState.Ready
 import edu.uci.ics.amber.engine.common.virtualidentity.LinkIdentity
+import edu.uci.ics.amber.engine.common.worker.WorkerState.Ready
 
 object AddPartitioningHandler {
   final case class AddPartitioning(tag: LinkIdentity, partitioning: Partitioning)
